requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
autoflake==0.6.6
beautifulsoup4==4.5.1
bs4==0.0.1
cachetools==2.0.1
cffi==1.11.2
chardet==3.0.4
Cogs==0.4.1
colour==0.1.2
conda==4.2.11
contextlib2==0.5.5
decorator==4.1.2
discord.py==0.16.12
emoji==0.4.5
Events==0.2.2
gevent==1.2.1
graphviz==0.6
greenlet==0.4.12
grequests==0.3.0
guess-language-spirit==0.5.3
imgurpython==1.1.7
jsonpickle==0.9.3
language-check==1.0
multidict==3.3.2
numpy==1.13.3
oauthlib==2.0.2
objgraph==3.1.0
Pillow==3.4.2
py-trello==0.9.0
pycosat==0.6.1
pycparser==2.17
pycrypto==2.6.1
pydot==1.2.3
pyenchant==1.6.8
pyflakes==1.5.0
pygame==1.9.2
pymongo==3.4.0
Pympler==0.5
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
raven==6.1.0
repoze.timeago==0.5
requests==2.11.1
requests-oauthlib==0.8.0
ruamel-yaml===-VERSION
six==1.10.0
ssdeep==3.2
validators==0.12.0
websockets==3.4
xdot==0.7
